# packages/server/src/reindex_server/openapi/reindex-http-v1.yaml
openapi: 3.1.0
info:
  title: ReIndex API
  description: 'ReIndex 1.0 package storage, versioning, retrieval, and search API.


    The HTTP v1 contract is backward compatible within the v1 path namespace. The

    three-stage publish flow is start upload, upload missing blobs, then commit.

    '
  version: 0.2.0
paths:
...
  /v1/push:
    post:
      tags:
      - Versions
      summary: Push
      operationId: push_v1_push_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushStartResponse'
        '400':
          description: Invalid request or cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection or resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Collection or model state conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request schema validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/push/blob:
    post:
      tags:
      - Versions
      summary: Push Blob
      operationId: push_blob_v1_push_blob_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_push_blob_v1_push_blob_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlobUploadResponse'
        '400':
          description: Invalid request or cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection or resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Collection or model state conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request schema validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/push/commit:
    post:
      tags:
      - Versions
      summary: Commit Push
      operationId: commit_push_v1_push_commit_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionedPushResponse'
        '400':
          description: Invalid request or cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection or resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Collection or model state conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request schema validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/fetch:
    post:
      tags:
      - Versions
      summary: Fetch
      operationId: fetch_v1_fetch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FetchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchResponse'
        '400':
          description: Invalid request or cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection or resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Collection or model state conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request schema validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/history:
    post:
      tags:
      - Versions
      summary: History
      operationId: history_v1_history_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
        '400':
          description: Invalid request or cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection or resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Collection or model state conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request schema validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/pull:
    post:
      tags:
      - Versions
      summary: Pull
      operationId: pull_v1_pull_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedCollectionRequest'
        required: true
      responses:
        '200':
          description: Node-only ZIP for the selected Collection version.
          headers:
            Content-Disposition:
              schema:
                type: string
            Content-Length:
              schema:
                type: string
            X-ReIndex-Package-Hash:
              schema:
                type: string
            X-ReIndex-Version-ID:
              schema:
                type: string
                format: uuid
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid request or cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection or resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Collection or model state conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request schema validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/search:
    post:
      tags:
      - Search
      summary: Search
      operationId: search_v1_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchApiResponse'
        '400':
          description: Invalid request or cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection or resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Collection or model state conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request schema validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/grep:
    post:
      tags:
      - Search
      summary: Grep
      operationId: grep_v1_grep_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrepRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchApiResponse'
        '400':
          description: Invalid request or cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection or resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Collection or model state conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request schema validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tables/query:
    post:
      tags:
      - Search
      summary: Query Table
      operationId: query_table_v1_tables_query_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableQueryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableQueryResponse'
        '400':
          description: Invalid request or cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection or resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Collection or model state conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request schema validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    BlobUploadResponse:
      properties:
        status:
          type: string
          enum:
          - stored
          - reused
          title: Status
        sha256:
          type: string
          title: Sha256
        byte_size:
          type: integer
          title: Byte Size
      type: object
      required:
      - status
      - sha256
      - byte_size
      title: BlobUploadResponse
    Body_push_blob_v1_push_blob_post:
      properties:
        upload_id:
          type: string
          title: Upload Id
        sha256:
          type: string
          title: Sha256
        blob:
          type: string
          contentMediaType: application/octet-stream
          title: Blob
      type: object
      required:
      - upload_id
      - sha256
      - blob
      title: Body_push_blob_v1_push_blob_post
...
    CommitRequest:
      properties:
        upload_id:
          type: string
          format: uuid
          title: Upload Id
      additionalProperties: false
      type: object
      required:
      - upload_id
      title: CommitRequest
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        request_id:
          type: string
          title: Request Id
        details:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Details
      type: object
      required:
      - code
      - message
      - request_id
      title: ErrorDetail
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - error
      title: ErrorResponse
...
    FetchRequest:
      properties:
        collection:
          type: string
          maxLength: 80
          minLength: 1
          title: Collection
          description: User-facing Collection name.
        version_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Version Id
      additionalProperties: false
      type: object
      required:
      - collection
      title: FetchRequest
    FetchResponse:
      properties:
        name:
          type: string
          title: Name
        collection_id:
          type: string
          title: Collection Id
        version:
          additionalProperties: true
          type: object
          title: Version
        manifest:
          additionalProperties: true
          type: object
          title: Manifest
      type: object
      required:
      - name
      - collection_id
      - version
      - manifest
      title: FetchResponse
...
    GrepRequest:
      properties:
        collection:
          type: string
          maxLength: 80
          minLength: 1
          title: Collection
          description: User-facing Collection name.
        pattern:
          type: string
          maxLength: 256
          minLength: 1
          title: Pattern
          description: Literal text or PostgreSQL regex.
        regex:
          type: boolean
          title: Regex
          description: Interpret pattern as a regex.
          default: false
        case_sensitive:
          type: boolean
          title: Case Sensitive
          description: Use case-sensitive matching.
          default: false
        limit:
          type: integer
          maximum: 50.0
          minimum: 1.0
          title: Limit
          description: Maximum grep hits.
          default: 10
      additionalProperties: false
      type: object
      required:
      - collection
      - pattern
      title: GrepRequest
...
    HistoryRequest:
      properties:
        collection:
          type: string
          maxLength: 80
          minLength: 1
          title: Collection
          description: User-facing Collection name.
        limit:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Limit
          default: 20
        cursor:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Cursor
      additionalProperties: false
      type: object
      required:
      - collection
      title: HistoryRequest
    HistoryResponse:
      properties:
        name:
          type: string
          title: Name
        collection_id:
          type: string
          title: Collection Id
        versions:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Versions
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      type: object
      required:
      - name
      - collection_id
      - versions
      title: HistoryResponse
...
    PushRequest:
      properties:
        name:
          type: string
          maxLength: 80
          minLength: 1
          title: Name
        collection_id:
          type: string
          format: uuid
          title: Collection Id
        base_version_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Base Version Id
        message:
          type: string
          maxLength: 1000
          minLength: 1
          title: Message
          default: Publish Collection
        operation:
          type: string
          enum:
          - publish
          - rollback
          title: Operation
          default: publish
        source_version_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Source Version Id
        dry_run:
          type: boolean
          title: Dry Run
          default: false
        manifest:
          $ref: '#/components/schemas/TransportManifest'
      additionalProperties: false
      type: object
      required:
      - name
      - collection_id
      - manifest
      title: PushRequest
    PushStartResponse:
      properties:
        status:
          type: string
          enum:
          - upload
          - ready
          - planned
          title: Status
        upload_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Upload Id
        expires_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Expires At
        head_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Head Version Id
        package_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Package Hash
        missing_blobs:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Missing Blobs
        no_op:
          type: boolean
          title: No Op
      type: object
      required:
      - status
      - missing_blobs
      - no_op
      title: PushStartResponse
    SearchApiResponse:
      properties:
        executed_mode:
          type: string
          enum:
          - lexical
          - semantic
          - hybrid
          - grep
          title: Executed Mode
        embedding_profile:
          anyOf:
          - type: string
          - type: 'null'
          title: Embedding Profile
        candidate_count:
          type: integer
          minimum: 0.0
          title: Candidate Count
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        results:
          items:
            $ref: '#/components/schemas/SearchResult'
          type: array
          title: Results
        applied:
          anyOf:
          - $ref: '#/components/schemas/AppliedSearch'
          - type: 'null'
        reranking:
          anyOf:
          - $ref: '#/components/schemas/AppliedReranking'
          - type: 'null'
      type: object
      required:
      - executed_mode
      - embedding_profile
      - candidate_count
      - next_cursor
      - results
      title: SearchApiResponse
    SearchFilters:
      properties:
        node_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 100
          title: Node Ids
          description: Restrict retrieval to these Node UUIDs.
        kinds:
          items:
            type: string
            enum:
            - group
            - text
            - table
            - image
            - file
          type: array
          maxItems: 5
          title: Kinds
          description: Restrict retrieval to these Node kinds.
        path_prefix:
          anyOf:
          - type: string
            maxLength: 500
            minLength: 1
          - type: 'null'
          title: Path Prefix
          description: Restrict retrieval to Node paths with this prefix.
        subtree_node_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Subtree Node Id
          description: Restrict retrieval to this Node subtree.
      additionalProperties: false
      type: object
      title: SearchFilters
...
    SearchRequest:
      properties:
        collection:
          type: string
          maxLength: 80
          minLength: 1
          title: Collection
          description: User-facing Collection name.
        query:
          type: string
          maxLength: 1000
          minLength: 1
          title: Query
          description: Natural-language search query.
        mode:
          type: string
          enum:
          - lexical
          - semantic
          - hybrid
          title: Mode
          description: Retrieval channels to execute.
          default: hybrid
        limit:
          type: integer
          maximum: 50.0
          minimum: 1.0
          title: Limit
          description: Maximum results in this page.
          default: 10
        candidate_limit:
          type: integer
          maximum: 500.0
          minimum: 10.0
          title: Candidate Limit
          description: Candidates retrieved per active channel before fusion.
          default: 100
        cursor:
          anyOf:
          - type: string
            maxLength: 2048
            minLength: 1
          - type: 'null'
          title: Cursor
          description: Opaque next_cursor returned by the previous identical search.
        filters:
          $ref: '#/components/schemas/SearchFilters'
        ranking:
          $ref: '#/components/schemas/SearchRanking'
      additionalProperties: false
      type: object
      required:
      - collection
      - query
      title: SearchRequest
      examples:
      - candidate_limit: 100
        collection: energy-reports
        filters:
          kinds:
          - text
          - table
        limit: 10
        mode: hybrid
        query: 未来十年光伏装机容量会增长到多少？
        ranking:
          lexical_weight: 0.5
          max_per_node: 3
          rrf_k: 60
          semantic_weight: 1.0
...
    TableQueryRequest:
      properties:
        collection:
          type: string
          maxLength: 80
          minLength: 1
          title: Collection
          description: User-facing Collection name.
        node_id:
          type: string
          format: uuid
          title: Node Id
          description: Node UUID in the current Collection state.
        sql:
          type: string
          maxLength: 10000
          minLength: 1
          title: Sql
        params:
          items:
            anyOf:
            - type: string
            - type: integer
            - type: number
            - type: boolean
            - type: 'null'
          type: array
          title: Params
      additionalProperties: false
      type: object
      required:
      - collection
      - node_id
      - sql
      title: TableQueryRequest
    TableQueryResponse:
      properties:
        columns:
          items:
            type: string
          type: array
          title: Columns
        rows:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Rows
        truncated:
          type: boolean
          title: Truncated
      type: object
      required:
      - columns
      - rows
      - truncated
      title: TableQueryResponse
...
    TransportManifest:
      properties:
        spec:
          type: string
          const: reindex/transport@1.0
          title: Spec
        package_root:
          type: string
          maxLength: 255
          minLength: 1
          title: Package Root
        files:
          items:
            $ref: '#/components/schemas/TransportFile'
          type: array
          maxItems: 100000
          minItems: 1
          title: Files
      additionalProperties: false
      type: object
      required:
      - spec
      - package_root
      - files
      title: TransportManifest
    VersionedCollectionRequest:
      properties:
        collection:
          type: string
          maxLength: 80
          minLength: 1
          title: Collection
          description: User-facing Collection name.
        version_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Version Id
      additionalProperties: false
      type: object
      required:
      - collection
      title: VersionedCollectionRequest
    VersionedPushResponse:
      properties:
        status:
          type: string
          const: ready
          title: Status
        name:
          type: string
          title: Name
        collection_id:
          type: string
          title: Collection Id
        package_hash:
          type: string
          title: Package Hash
        nodes:
          type: integer
          title: Nodes
        sources:
          type: integer
          title: Sources
        resources:
          type: integer
          title: Resources
        search_units:
          type: integer
          title: Search Units
        embedding_profile:
          anyOf:
          - type: string
          - type: 'null'
          title: Embedding Profile
        version_id:
          type: string
          title: Version Id
        parent_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Version Id
        operation:
          type: string
          enum:
          - publish
          - rollback
          title: Operation
        source_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Version Id
        uploaded_blobs:
          type: integer
          title: Uploaded Blobs
        reused_blobs:
          type: integer
          title: Reused Blobs
        embedded_units:
          type: integer
          title: Embedded Units
        reused_embeddings:
          type: integer
          title: Reused Embeddings
        no_op:
          type: boolean
          title: No Op
          default: false
      type: object
      required:
      - status
      - name
      - collection_id
      - package_hash
      - nodes
      - sources
      - resources
      - search_units
      - embedding_profile
      - version_id
      - operation
      - uploaded_blobs
      - reused_blobs
      - embedded_units
      - reused_embeddings
      title: VersionedPushResponse
...